NCBC-2860: Configuration revisions should be parsed and compared with 64-bit precision
Motivation
----------
Align .NET SDK with other SDKs with respect to supporting unsigned 64
bit revision numbers.

Modifications
-------------
 - Replace all uint representing the revision with ulong

Result
------
Revision is stored as a 64 but unsigned integer.

Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/151581
Tested-by: Build Bot <build@couchbase.com>

+++ src/Couchbase/Core/Configuration/Server/BucketConfig.cs
@@ -323,7 +323,7 @@ namespace Couchbase.Core.Configuration.Server
         /// </summary>
         [JsonIgnore] public bool IsGlobal { get; set; }
 
-        [JsonProperty("rev")] public uint Rev { get; set; }
+        [JsonProperty("rev")] public ulong Rev { get; set; }
         [JsonProperty("name")] public string Name { get; set; }
         [JsonProperty("uri")] public string Uri { get; set; }
         [JsonProperty("streamingUri")] public string StreamingUri { get; set; }
